backend/requirements.txt
opencv-python
numpy
Pillow
mediapipe
pytesseract
pyzbar
scikit-image